Major change! Display framerate is now a float!
Had to change sdl's sdl_video.h bullshit cuz it had int refresh rate and I had finally updated my fork of libsdl just to copy over changes from my lime fork's project\lib\sdl folder.

+++ project/include/system/DisplayMode.h
@@ -16,12 +16,12 @@ namespace lime {
 			hl_type* t;
 			int height;
 			PixelFormat pixelFormat;
-			int refreshRate;
+			double refreshRate;
 			int width;
 
 			DisplayMode ();
 			DisplayMode (value DisplayMode);
-			DisplayMode (int width, int height, PixelFormat pixelFormat, int refreshRate);
+			DisplayMode (int width, int height, PixelFormat pixelFormat, double refreshRate);
 
 			void CopyFrom (DisplayMode* other);
 			void* Value ();
